requirements.txt
mcp
httpx
google-auth
google-auth-oauthlib
google-api-python-client
